+++ regal-api/src/main/java/de/nrw/hbz/regal/api/Resource.java
@@ -33,6 +33,10 @@ import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
+import javax.ws.rs.core.CacheControl;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.EntityTag;
+import javax.ws.rs.core.Request;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.ResponseBuilder;
 import javax.ws.rs.core.Response.Status;
@@ -767,63 +771,131 @@ public class Resource {
     }
 
     /**
-     * Returns a text extraction, if the data is of mimetype application/pdf
+     * Returns a oai-dc conversion of pid's metadata
      * 
      * @param pid
-     *            the pid of the resource
-     * @param namespace
-     *            the namespace of the resource
-     * @return a text/plain representation of the pdf contente, if any.
+     *            the metadata of the identified resource will be transformed to
+     *            oaidc
+     * @return the oai_dc xml
      */
     @GET
-    @Path("/{namespace}:{pid}/fulltext")
-    @Produces({ "application/xml", "application/json" })
-    public Response getFulltext(@PathParam("pid") String pid,
-	    @PathParam("namespace") String namespace) {
+    @Path("/{pid}.oaidc")
+    @Produces({ "application/xml" })
+    public String oaidc(@PathParam("pid") String pid) {
 	try {
-	    return actions.getFulltext(pid, namespace);
-	} catch (URISyntaxException e) {
+	    return actions.oaidc(pid);
+	} catch (ArchiveException e) {
 	    throw new HttpArchiveException(
 		    Status.INTERNAL_SERVER_ERROR.getStatusCode(), e);
 	}
-
     }
 
     /**
-     * Returns a xepicur stream for the resource
+     * Returns epicur for the pid, if urn is available
      * 
      * @param pid
-     *            the pid of the resource
-     * @param namespace
-     *            the namespace of the resource
-     * @return an epicur entry for the resource
+     *            pid with namespace
+     * @return epicur xml
      */
     @GET
-    @Path("/{namespace}:{pid}.epicur")
-    @Produces({ "application/xml" })
-    public String getEpicur(@PathParam("pid") String pid,
-	    @PathParam("namespace") String namespace) {
-
-	return actions.epicur(pid, namespace);
-
+    @Path("/{pid}.epicur")
+    @Produces({ "application/json", "application/xml" })
+    public String epicur(@PathParam("pid") String pid) {
+	try {
+	    return actions.epicur(pid);
+	} catch (ArchiveException e) {
+	    throw new HttpArchiveException(
+		    Status.INTERNAL_SERVER_ERROR.getStatusCode(), e);
+	}
     }
 
     /**
-     * Returns a oai-dc representation converted from the metadata
+     * Extractes text from pid/data if data exists and is a pdf
      * 
      * @param pid
-     *            the pid of the resource the pid of the resource
-     * @param namespace
-     *            the namespace of the resource
-     * @return an aggregated representation of the resource
+     *            the pid must contain a data stream with mime type
+     *            application/pdf
+     * @param request
+     *            lastModified is checked.
+     * @return a text/plain message containing the extracted text
      */
     @GET
-    @Path("/{namespace}:{pid}.oaidc")
-    @Produces({ "application/xml" })
-    public String getDC(@PathParam("pid") String pid,
-	    @PathParam("namespace") String namespace) {
+    @Path("/{pid}.pdfbox")
+    @Produces({ "text/plain; charset=UTF-8" })
+    public Response pdfbox(@PathParam("pid") String pid,
+	    @Context Request request) {
+	try {
+	    Node node = actions.readNode(pid);
 
-	return actions.oaidc(namespace + ":" + pid);
+	    final EntityTag eTag = new EntityTag(node.getPID() + "_"
+		    + node.getLastModified().getTime());
 
+	    final CacheControl cacheControl = new CacheControl();
+	    cacheControl.setMaxAge(-1);
+
+	    ResponseBuilder builder = request.evaluatePreconditions(
+		    node.getLastModified(), eTag);
+
+	    // the user's information was modified, return it
+	    if (builder == null) {
+		builder = Response.ok(actions.pdfbox(node));
+	    }
+
+	    // the user's information was not modified, return a 304
+	    return builder.cacheControl(cacheControl)
+		    .lastModified(node.getLastModified()).tag(eTag).build();
+
+	} catch (ArchiveException e) {
+	    throw new HttpArchiveException(
+		    Status.INTERNAL_SERVER_ERROR.getStatusCode(), e);
+	}
+    }
+
+    /**
+     * Convertes pid/data to pdfA if data exists and is a pdf
+     * 
+     * @param pid
+     *            the pid must contain a data stream with mime type
+     *            application/pdf
+     * @param request
+     *            lastModified is checked.
+     * @return a text/plain message containing the extracted text
+     */
+    @GET
+    @Path("/{pid}.pdfa")
+    @Produces({ "text/plain; charset=UTF-8" })
+    public Response pdfa(@PathParam("pid") String pid, @Context Request request) {
+	try {
+	    Node node = actions.readNode(pid);
+
+	    final EntityTag eTag = new EntityTag(node.getPID() + "_"
+		    + node.getLastModified().getTime());
+
+	    final CacheControl cacheControl = new CacheControl();
+	    cacheControl.setMaxAge(-1);
+
+	    ResponseBuilder builder = request.evaluatePreconditions(
+		    node.getLastModified(), eTag);
+
+	    // the user's information was modified, return it
+	    if (builder == null) {
+		String redirectUrl = actions.pdfa(node);
+		try {
+		    builder = Response.temporaryRedirect(
+			    new java.net.URI(redirectUrl)).status(303);
+		} catch (URISyntaxException e) {
+		    throw new HttpArchiveException(
+			    Status.INTERNAL_SERVER_ERROR.getStatusCode(), e);
+		}
+	    }
+
+	    // the user's information was not modified, return a 304
+	    return builder.cacheControl(cacheControl)
+		    .lastModified(node.getLastModified()).tag(eTag).build();
+
+	} catch (ArchiveException e) {
+	    throw new HttpArchiveException(
+		    Status.INTERNAL_SERVER_ERROR.getStatusCode(), e);
+	}
     }
 }
